Convert loops using unsigned integers as iterators to use regular signed ints instead
https://kristerw.blogspot.com/2016/02/how-undefined-signed-overflow-enables.html

Doing this as cleanly as possible involved demoting several function parameters concerning object sizes and counts from size_t to int--I'm fine with this change as the functions in question are not actually capable of handling input with sizes larger than what can be stored in a signed 32-bit integer, making the use of size_t here misleading at best.

git-svn-id: https://svn.eduke32.com/eduke32@7673 1a8010ca-5511-0410-912e-c29ae57300e0

# Conflicts:
#	source/build/src/polymost.cpp
#	source/build/src/texcache.cpp
#	source/build/src/tilepacker.cpp

+++ source/audiolib/src/flac.cpp
@@ -547,21 +547,21 @@ int32_t MV_PlayFLAC(char *ptr, uint32_t length, int32_t loopstart, int32_t loope
                                 const size_t field = value - entry;
                                 value += 1;
 
-                                for (size_t t = 0; t < loopStartTagCount && vc_loopstart == NULL; ++t)
+                                for (int t = 0; t < loopStartTagCount && vc_loopstart == NULL; ++t)
                                 {
                                     char const * const tag = loopStartTags[t];
                                     if (field == strlen(tag) && Bstrncasecmp(entry, tag, field) == 0)
                                         vc_loopstart = Xstrdup(value);
                                 }
 
-                                for (size_t t = 0; t < loopEndTagCount && vc_loopend == NULL; ++t)
+                                for (int t = 0; t < loopEndTagCount && vc_loopend == NULL; ++t)
                                 {
                                     char const * const tag = loopEndTags[t];
                                     if (field == strlen(tag) && Bstrncasecmp(entry, tag, field) == 0)
                                         vc_loopend = Xstrdup(value);
                                 }
 
-                                for (size_t t = 0; t < loopLengthTagCount && vc_looplength == NULL; ++t)
+                                for (int t = 0; t < loopLengthTagCount && vc_looplength == NULL; ++t)
                                 {
                                     char const * const tag = loopLengthTags[t];
                                     if (field == strlen(tag) && Bstrncasecmp(entry, tag, field) == 0)
